+++ rsx_decompiler/rsx_vp_decompiler.cpp
@@ -193,15 +193,15 @@ namespace rsx
 
 				auto get_register = [&]()
 				{
-					switch (src.register_type)
+					switch ((u32)src.register_type)
 					{
-					case src_register_type::temporary:
+					case (u32)src_register_type::temporary:
 						return context.temporary(src.tmp_src);
 
-					case src_register_type::input:
+					case (u32)src_register_type::input:
 						return context.input(instruction.data.d1.input_src);
 
-					case src_register_type::constant:
+					case (u32)src_register_type::constant:
 						return context.constant(instruction.data.d1.const_src,
 							instruction.data.d3.index_const,
 							instruction.data.d0.addr_reg_sel_1,
@@ -533,15 +533,15 @@ namespace rsx
 			{
 				is_vec = false;
 
-				switch (instruction.data.d1.sca_opcode)
+				switch ((u32)instruction.data.d1.sca_opcode)
 				{
-				case sca_opcode::mov: return set_dst(src_swizzled_as_dst(2));
-				case sca_opcode::rcp: return set_dst(float_point_t<1>::ctor(1.0f) / src_swizzled_as_dst(2));
-				case sca_opcode::rcc: return set_dst(base::clamp(float_point_t<1>::ctor(1.0f) / src_swizzled_as_dst(2), 5.42101e-20f, 1.884467e19f));
-				case sca_opcode::rsq: return set_dst(base::rsqrt(base::abs(src_swizzled_as_dst(2))));
-				case sca_opcode::exp: return set_dst(base::exp(src_swizzled_as_dst(2)));
-				case sca_opcode::log: return set_dst(base::log(src_swizzled_as_dst(2)));
-				case sca_opcode::lit:
+				case (u32)sca_opcode::mov: return set_dst(src_swizzled_as_dst(2));
+				case (u32)sca_opcode::rcp: return set_dst(float_point_t<1>::ctor(1.0f) / src_swizzled_as_dst(2));
+				case (u32)sca_opcode::rcc: return set_dst(base::clamp(float_point_t<1>::ctor(1.0f) / src_swizzled_as_dst(2), 5.42101e-20f, 1.884467e19f));
+				case (u32)sca_opcode::rsq: return set_dst(base::rsqrt(base::abs(src_swizzled_as_dst(2))));
+				case (u32)sca_opcode::exp: return set_dst(base::exp(src_swizzled_as_dst(2)));
+				case (u32)sca_opcode::log: return set_dst(base::log(src_swizzled_as_dst(2)));
+				case (u32)sca_opcode::lit:
 				{
 					auto t = src(2);
 
@@ -550,8 +550,8 @@ namespace rsx
 
 					return set_dst(swizzle_as_dst(float_point_t<4>::ctor(1.0f, t.x(), z_value, 1.0f)));
 				}
-				case sca_opcode::bra: break;
-				case sca_opcode::bri:
+				case (u32)sca_opcode::bra: break;
+				case (u32)sca_opcode::bri:
 				{
 					std::size_t from = base::writer.position;
 					std::size_t to = address_value();
@@ -569,17 +569,17 @@ namespace rsx
 				}
 				return{ "" };
 
-				case sca_opcode::cal: break;
-				case sca_opcode::cli: break;
-				case sca_opcode::ret: return conditional(typename base::void_expr{ "return" });
-				case sca_opcode::lg2: return set_dst(base::log2(src_swizzled_as_dst(2)));
-				case sca_opcode::ex2: return set_dst(base::exp2(src_swizzled_as_dst(2)));
-				case sca_opcode::sin: return set_dst(base::sin(src_swizzled_as_dst(2)));
-				case sca_opcode::cos: return set_dst(base::cos(src_swizzled_as_dst(2)));
-				case sca_opcode::brb: break;
-				case sca_opcode::clb: break;
-				case sca_opcode::psh: break;
-				case sca_opcode::pop: break;
+				case (u32)sca_opcode::cal: break;
+				case (u32)sca_opcode::cli: break;
+				case (u32)sca_opcode::ret: return conditional(typename base::void_expr{ "return" });
+				case (u32)sca_opcode::lg2: return set_dst(base::log2(src_swizzled_as_dst(2)));
+				case (u32)sca_opcode::ex2: return set_dst(base::exp2(src_swizzled_as_dst(2)));
+				case (u32)sca_opcode::sin: return set_dst(base::sin(src_swizzled_as_dst(2)));
+				case (u32)sca_opcode::cos: return set_dst(base::cos(src_swizzled_as_dst(2)));
+				case (u32)sca_opcode::brb: break;
+				case (u32)sca_opcode::clb: break;
+				case (u32)sca_opcode::psh: break;
+				case (u32)sca_opcode::pop: break;
 				default:
 					throw;
 				}
@@ -591,31 +591,31 @@ namespace rsx
 			{
 				is_vec = true;
 
-				switch (instruction.data.d1.vec_opcode)
+				switch ((u32)instruction.data.d1.vec_opcode)
 				{
-				case vec_opcode::mov: return set_dst(src_swizzled_as_dst(0));
-				case vec_opcode::mul: return set_dst(src_swizzled_as_dst(0) * src_swizzled_as_dst(1));
-				case vec_opcode::add: return set_dst(src_swizzled_as_dst(0) + src_swizzled_as_dst(2));
-				case vec_opcode::mad: return set_dst((src_swizzled_as_dst(0) * src_swizzled_as_dst(1)).without_scope() + src_swizzled_as_dst(2));
-				case vec_opcode::dp3: return set_dst(base::dot(src(0).xyz(), src(1).xyz()));
-				case vec_opcode::dph: break;
-				case vec_opcode::dp4: return set_dst(base::dot(src(0), src(1)));
-				case vec_opcode::dst: break;
-				case vec_opcode::min: return set_dst(base::min(src_swizzled_as_dst(0), src_swizzled_as_dst(1)));
-				case vec_opcode::max: return set_dst(base::max(src_swizzled_as_dst(0), src_swizzled_as_dst(1)));
-				case vec_opcode::slt: return set_dst(compare(base::compare_function::less, src_swizzled_as_dst(0), src_swizzled_as_dst(1)));
-				case vec_opcode::sge: return set_dst(compare(base::compare_function::greater_equal, src_swizzled_as_dst(0), src_swizzled_as_dst(1)));
-				case vec_opcode::arl: return typename base::writer_t{} += address_register() = integer_t<1>::ctor(src(0).x());
-				case vec_opcode::frc: return set_dst(base::fract(src_swizzled_as_dst(0)));
-				case vec_opcode::flr: return set_dst(base::floor(src_swizzled_as_dst(0)));;
-				case vec_opcode::seq: return set_dst(compare(base::compare_function::equal, src_swizzled_as_dst(0), src_swizzled_as_dst(1)));
-				case vec_opcode::sfl: return set_dst(0.0f);
-				case vec_opcode::sgt: return set_dst(compare(base::compare_function::greater, src_swizzled_as_dst(0), src_swizzled_as_dst(1)));;
-				case vec_opcode::sle: return set_dst(compare(base::compare_function::less_equal, src_swizzled_as_dst(0), src_swizzled_as_dst(1)));
-				case vec_opcode::sne: return set_dst(compare(base::compare_function::not_equal, src_swizzled_as_dst(0), src_swizzled_as_dst(1)));
-				case vec_opcode::str: return set_dst(1.0f);
-				case vec_opcode::ssg: break;
-				case vec_opcode::txl: break;
+				case (u32)vec_opcode::mov: return set_dst(src_swizzled_as_dst(0));
+				case (u32)vec_opcode::mul: return set_dst(src_swizzled_as_dst(0) * src_swizzled_as_dst(1));
+				case (u32)vec_opcode::add: return set_dst(src_swizzled_as_dst(0) + src_swizzled_as_dst(2));
+				case (u32)vec_opcode::mad: return set_dst((src_swizzled_as_dst(0) * src_swizzled_as_dst(1)).without_scope() + src_swizzled_as_dst(2));
+				case (u32)vec_opcode::dp3: return set_dst(base::dot(src(0).xyz(), src(1).xyz()));
+				case (u32)vec_opcode::dph: break;
+				case (u32)vec_opcode::dp4: return set_dst(base::dot(src(0), src(1)));
+				case (u32)vec_opcode::dst: break;
+				case (u32)vec_opcode::min: return set_dst(base::min(src_swizzled_as_dst(0), src_swizzled_as_dst(1)));
+				case (u32)vec_opcode::max: return set_dst(base::max(src_swizzled_as_dst(0), src_swizzled_as_dst(1)));
+				case (u32)vec_opcode::slt: return set_dst(compare(base::compare_function::less, src_swizzled_as_dst(0), src_swizzled_as_dst(1)));
+				case (u32)vec_opcode::sge: return set_dst(compare(base::compare_function::greater_equal, src_swizzled_as_dst(0), src_swizzled_as_dst(1)));
+				case (u32)vec_opcode::arl: return typename base::writer_t{} += address_register() = integer_t<1>::ctor(src(0).x());
+				case (u32)vec_opcode::frc: return set_dst(base::fract(src_swizzled_as_dst(0)));
+				case (u32)vec_opcode::flr: return set_dst(base::floor(src_swizzled_as_dst(0)));;
+				case (u32)vec_opcode::seq: return set_dst(compare(base::compare_function::equal, src_swizzled_as_dst(0), src_swizzled_as_dst(1)));
+				case (u32)vec_opcode::sfl: return set_dst(0.0f);
+				case (u32)vec_opcode::sgt: return set_dst(compare(base::compare_function::greater, src_swizzled_as_dst(0), src_swizzled_as_dst(1)));;
+				case (u32)vec_opcode::sle: return set_dst(compare(base::compare_function::less_equal, src_swizzled_as_dst(0), src_swizzled_as_dst(1)));
+				case (u32)vec_opcode::sne: return set_dst(compare(base::compare_function::not_equal, src_swizzled_as_dst(0), src_swizzled_as_dst(1)));
+				case (u32)vec_opcode::str: return set_dst(1.0f);
+				case (u32)vec_opcode::ssg: break;
+				case (u32)vec_opcode::txl: break;
 
 				default:
 					throw;
